Updated to telebot v3

+++ internal/domain/transaction.go
@@ -5,7 +5,7 @@ import "time"
 type TransactionDTO struct {
 	ID          string     `json:"id,omitempty"`
 	MsgID       int        `json:"msg_id,omitempty"`
-	UserID      int        `json:"user_id,omitempty"`
+	UserID      int64      `json:"user_id,omitempty"`
 	Amount      float64    `json:"amount"`
 	Description string     `json:"description"`
 	WalletID    int        `json:"wallet_id"`
